feat: [온보딩] [정보-학교-캐릭트-라이프스타일-선호칩] 순서 반영
