+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Cozymate"
         tools:targetApi="31">
+        <activity android:name=".ui.onboarding.OnboardingSelectingFragmentActivity"
+            android:exported="false" />
         <activity
             android:name=".ui.my_page.update_room.UpdateRoomInfoActivity"
             android:exported="false" />
