+++ app/src/main/java/umc/cozymate/ui/roommate/RoommateInputInfoActivity.kt
@@ -19,6 +19,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import umc.cozymate.data.model.request.UserInfoRequest
 import umc.cozymate.databinding.ActivityRoommateInputInfoBinding
 import umc.cozymate.ui.MainActivity
+import umc.cozymate.ui.onboarding.OnboardingActivity
+import umc.cozymate.ui.onboarding.OnboardingSelectingFragmentActivity
 import umc.cozymate.ui.roommate.adapter.RoommateInputInfoVPA
 import umc.cozymate.ui.roommate.lifestyle_info.BasicInfoFragment
 import umc.cozymate.ui.roommate.lifestyle_info.EssentialInfoFragment
@@ -68,9 +70,9 @@ class RoommateInputInfoActivity : AppCompatActivity() {
                 sendUserDataToViewModel()
                 Log.d("RoommateInputInfoActivity", "sendUserInfo")
 
-                // 마지막 프래그먼트에서 메인 액티비티로 이동
-                val intent = Intent(this, MainActivity::class.java)
-                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
+                // 마지막 프래그먼트에서 온보딩 액티비티로 이동
+                val intent = Intent(this, OnboardingSelectingFragmentActivity::class.java)
+                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
                 startActivity(intent)
                 finish()
             }
@@ -198,6 +200,6 @@ class RoommateInputInfoActivity : AppCompatActivity() {
             selfIntroduction = spf.getString("user_selfIntroduction", "") ?: ""
         )
         viewModel.sendUserInfo(userInfo)
-        finish()
+        finish() // 결과 전달 후 종료
     }
 }
